requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blis==0.7.4
boto3==1.17.9
botocore==1.20.9
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
Cython==0.29.14
DAWG-Python==0.7.2
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flatbuffers==1.12
gast==0.3.3
gensim==3.8.3
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.7.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
lemminflect==0.2.2
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pathy==0.3.5
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.3
pylint==2.6.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.0.0/ru_core_news_sm-3.0.0-py3-none-any.whl
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.95
six==1.15.0
smart-open==3.0.0
spacy==3.0.1
spacy-legacy==3.0.1
spacy-udpipe==0.3.2
srsly==2.4.0
stanza==1.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
thinc==8.0.1
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.7.1+cu110
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
transformers==4.4.2
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
ufal.udpipe==1.2.0.3
urllib3==1.26.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
zipp==3.4.0
